evaluate {glue}	R Documentation

Evaluate R code

Description

This is a simple wrapper around eval(parse()) which provides a more consistent interface than the default functions. If data is NULL then the code is evaluated in the environment. If data is not NULL than the code is evaluated in the data object first, with the enclosing environment of envir.

Usage

evaluate(code, envir)

Arguments

code
R code to evaluate

envir
environment to evaluate the code in

Details

This function is designed to be used within transformers to evaluate the code in the glue block.

[Package glue version 1.2.0 Index]