glue-deprecated {glue}	R Documentation

Deprecated Functions

Description

These functions are Deprecated in this release of glue, they will be removed in a future version.

[Package glue version 1.6.2 Index]