quoting {glue}	R Documentation

Quoting operators

Description

These functions make it easy to quote each individual element and are useful in conjunction with glue_collapse().

Usage

single_quote(x)

double_quote(x)

backtick(x)

Arguments

x
A character to quote.

Examples

x <- 1:5
glue('Values of x: {glue_collapse(backtick(x), sep = ", ", last = " and ")}')

[Package glue version 1.6.1 Index]